refactor: 물리 엔진 관련 도미노 렌더링 로직 DominoEntity 컴포넌트로 분리
- threejs 관련 로직을 DominoCanvas컴포넌트로 분리하여 DominoHUD와 분리
- 도미노 렌더링 로직을 DominoEntity 컴포넌트로 분리하여 훅 사용 규칙 충돌 해결
- Canvas > Physics > DominoEntity 계층 구조로 명확한 책임 분리

[REFACTOR] 프론트엔드 전반 리팩토링 (렌더링 상태 분리, 매직넘버 제거 등) #74
